AI_Suggestion_App/backend/Groq_3_Suggest_llama3-8b/requirements.txt
flask==3.0.3
flask-cors==5.0.0
openai
python-dotenv
